+++ textblob/exceptions.py
@@ -11,22 +11,26 @@ or use the NLTK downloader to download the missing data: http://nltk.org/data.ht
 If this doesn't fix the problem, file an issue at https://github.com/sloria/TextBlob/issues.
 """
 
-class TextBlobException(Exception):
-    """A TextBlob-related exception."""
+class TextBlobError(Exception):
+    """A TextBlob-related error."""
     pass
 
-class MissingCorpusException(TextBlobException):
+TextBlobException = TextBlobError  # Backwards compat
+
+class MissingCorpusError(TextBlobError):
     """Exception thrown when a user tries to use a feature that requires a
     dataset or model that the user does not have on their system.
     """
 
     def __init__(self, message=MISSING_CORPUS_MESSAGE, *args, **kwargs):
-        super(MissingCorpusException, self).__init__(message, *args, **kwargs)
+        super(MissingCorpusError, self).__init__(message, *args, **kwargs)
 
-class DeprecationError(TextBlobException):
+MissingCorpusException = MissingCorpusError  # Backwards compat
+
+class DeprecationError(TextBlobError):
     """Raised when user uses a deprecated feature."""
     pass
 
-class TranslatorError(TextBlobException):
+class TranslatorError(TextBlobError):
     """Raised when an error occurs during language translation or detection."""
     pass